polish

+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: gozturk <marvin@codam.nl>                    +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/10/18 14:01:56 by gozturk       #+#    #+#                 */
-/*   Updated: 2022/11/03 14:56:41 by gozturk       ########   odam.nl         */
+/*   Updated: 2022/11/03 15:09:56 by gozturk       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ char	*ft_strjoin(char const *s1, char const *s2)
 		ptr[i] = s1[i];
 		i++;
 	}
-	while (i < s1_len + ft_strlen(s2) + 1)
+	while (i < (int)(s1_len + ft_strlen(s2) + 1))
 	{
 		ptr[i - 1] = s2[j];
 		i++;
